Move stock_investing_exception to domain
