+++ src/domain/stock_investing_exception.py
