+++ tests/unit/exceptions/test_stock_investing_exception.py
@@ -1,4 +1,4 @@
-from src.exceptions.stock_investing_exception import StockInvestingException
+from src.domain.stock_investing_exception import StockInvestingException
 
 
 def test_given_class_stock_investing_exception_when_initializing_then_code_and_message_are_none():
